functions/requirements.txt
torch==2.5.1
torchvision==0.20.1
pillow>=10.0.0
python-chess>=1.999 
CairoSVG==2.7.1
scikit-image==0.25.0
firebase-functions==0.4.2
functions-framework>=3.0.0